added a copy-pasta fabric.mod.json file
This is a file to edit and copy-pasta into the actual fabric.mod.json file when the mod is exported to a jar file. Unfortunately, MCreator doesn't allow customized fabric.mod.json files. I am looking into making a plugin that supports custom fabric.mod.json files for MCreator but haven't figured out how to do that yet. When I do, it will be required to clone this mod into MCreator.